+++ packages/components-react/src/pages/Modal.tsx
@@ -1,5 +1,5 @@
 /* Auto Generated File */
-import { PHeading, PModal, PorscheDesignSystemProvider, PText } from '@porsche-design-system/components-react';
+import { PHeading, PModal, PText } from '@porsche-design-system/components-react';
 
 export const ModalPage = (): JSX.Element => {
   const style = `
@@ -25,17 +25,11 @@ export const ModalPage = (): JSX.Element => {
   `;
 
   return (
-    <PorscheDesignSystemProvider prefix="my-prefix">
+    <>
       <style dangerouslySetInnerHTML={{ __html: style }} />
 
       <div className="playground light" title="should show modal">
         <PModal heading="Heading" open={true}>Some Content: should show modal</PModal>
-      </div>
-
-      <div className="playground light" title="should show modal with prefixed component initialization">
-        <PModal heading="Heading" open={true}>
-          Some Content: should show modal with prefixed component initialization
-        </PModal>
       </div>
 
       <div className="playground light" title="should show modal (fullscreen=true)">
@@ -158,6 +152,6 @@ export const ModalPage = (): JSX.Element => {
         <span />
         <span />
       </div>
-    </PorscheDesignSystemProvider>
+    </>
   );
 };