+++ packages/storefront/src/views/Page.vue
@@ -3,7 +3,6 @@
     <p-tabs-bar v-if="hasTabs" :active-tab-index="activeTabIndex" size="medium">
       <router-link v-for="(tab, index) in tabs" :key="index" :to="createTabLink(tab)">{{ tab }}</router-link>
     </p-tabs-bar>
-    <p-divider v-if="hasTabs"></p-divider>
     <Markdown>
       <component :is="component" v-for="(component, index) in components" :key="index"></component>
     </Markdown>
@@ -106,7 +105,9 @@
 </script>
 
 <style scoped lang="scss">
-  p-divider {
-    margin-bottom: 4rem;
+  @use '@porsche-design-system/components-js/styles' as *;
+
+  p-tabs-bar {
+    margin-bottom: $pds-spacing-fluid-large;
   }
 </style>