+++ packages/storefront/tests/e2e/specs/navigation.e2e.ts
@@ -18,8 +18,10 @@ beforeEach(async () => {
 });
 afterEach(async () => await browserPage.close());
 
+const isTabActive = (element: ElementHandle | null): Promise<boolean> =>
+  element.evaluate((el ) => el.className.includes('router-link-active'));
 const isLinkActive = (element: ElementHandle | null): Promise<boolean> =>
-  element.evaluate((el) => el.className.includes('router-link-active'));
+  element.evaluate((el ) => el.active);
 const getMainTitle = (page: Page): Promise<string> => page.$eval('.vmark > h1', (x) => x.innerHTML);
 const hasPageObjectObject = (page: Page): Promise<boolean> =>
   page.evaluate(() => document.body.innerText.includes('[object Object]'));
@@ -41,17 +43,17 @@ const injectCSSOverrides = async () => {
 // TODO: we shouldn't rely on retries since computed result has to be deterministic
 jest.retryTimes(1);
 
+// TODO: we should prepare a test array and loop by it.each instead to have better console output while test case is executed
 for (const [category, pages] of Object.entries(STOREFRONT_CONFIG)) {
   for (const [page, tabs] of Object.entries(pages).sort(([a], [b]) => a.localeCompare(b))) {
     ((category: string, page: string) => {
       it(`should navigate to "${category} > ${page}"`, async () => {
-        // console.log(`${category} > ${page}`);
         const [accordionButton] = await browserPage.$x(
-          `//div[@class='sidebar']/nav/p-accordion[@heading='${category}']`
+          `//aside//nav/p-accordion[@heading='${category}']`
         );
         const href = `\/${paramCase(category)}\/${paramCase(page)}`;
         const [linkElement] = await browserPage.$x(
-          `//div[@class='sidebar']/nav//p-link-pure//a[contains(., '${page}')][@href='${href}']`
+          `//aside//nav//p-link-pure/a[contains(., '${page}')][@href='${href}']/parent::p-link-pure`
         );
 
         await accordionButton.click();
@@ -73,7 +75,7 @@ for (const [category, pages] of Object.entries(STOREFRONT_CONFIG)) {
             const tabHref = `\/${paramCase(category)}\/${paramCase(page)}\/${paramCase(tab)}`;
             const [tabElement] = await browserPage.$x(`//p-tabs-bar//a[contains(., '${tab}')][@href='${tabHref}']`);
 
-            const isTabElementActiveInitially = await isLinkActive(tabElement);
+            const isTabElementActiveInitially = await isTabActive(tabElement);
             if (parseInt(index) === 0) {
               expect(isTabElementActiveInitially, 'should have first tab active initially').toBe(true);
             } else {
@@ -84,7 +86,7 @@ for (const [category, pages] of Object.entries(STOREFRONT_CONFIG)) {
             await browserPage.waitForNetworkIdle();
             await browserPage.evaluate(() => (window as any).componentsReady());
 
-            expect(await isLinkActive(tabElement), 'should have tab active after click').toBe(true);
+            expect(await isTabActive(tabElement), 'should have tab active after click').toBe(true);
             expect(await getMainTitle(browserPage), 'should show correct main title for tab page').toBe(page);
             expect(await hasPageObjectObject(browserPage), 'should not contain [object Object] on tab page').toBe(
               false