+++ packages/components/CHANGELOG.md
@@ -1,13 +1,45 @@
 # Changelog
 
-## Porsche Design System - Components
+## Porsche Design System
 
-All notable changes to this project will be documented in this file.
+All notable changes to this project will be documented in this file and published as following npm packages:
+
+- `@porsche-design-system/components-js`
+- `@porsche-design-system/components-angular`
+- `@porsche-design-system/components-react`
+- `@porsche-design-system/components-vue`
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### [Unreleased]
+
+#### Added
+
+- `skipLinkTarget` property to `Carousel` component to enhance keyboard functionality
+  ([#2557](https://github.com/porsche-design-system/porsche-design-system/pull/2557))
+
+#### Fixed
+
+- Partials: `getInitialStyles` supports `Flyout` component
+  ([#2598](https://github.com/porsche-design-system/porsche-design-system/pull/2598))
+- `Popover` content can be selected/highlighted
+  ([#2599](https://github.com/porsche-design-system/porsche-design-system/pull/2599))
+
+#### Changed
+
+- `Carousel` supports click events on non-active slides and changed keyboard navigation
+  ([#2557](https://github.com/porsche-design-system/porsche-design-system/pull/2557))
+
+### [3.1.0] - 2023-06-09
+
+### [3.1.0-rc.2] - 2023-06-09
+
+#### Changed
+
+- `Crest` updated assets ([#2595](https://github.com/porsche-design-system/porsche-design-system/pull/2595))
+- Partials: `getMetaTagsAndIconLinks` updated assets
+  ([#2595](https://github.com/porsche-design-system/porsche-design-system/pull/2595))
 
 #### Added